+++ src/sno_fo_fro/hypotheses.py
